requirements.txt
torch
torchvision
torchaudio
lightning
pyyaml
python-dotenv
numpy
pandas
matplotlib
seaborn
tqdm
Pillow
ipykernel
scikit-learn